Improvements to error messages and refactor files
